Try using vector paths

+++ src/laddermap/mod.rs
@@ -1,4 +1,5 @@
 use bevy::prelude::*;
+use bevy_prototype_lyon::prelude::*;
 
 use num_derive::FromPrimitive;
 #[derive(Copy, Clone, Default, Component, Debug, FromPrimitive)]
@@ -11,10 +12,10 @@ pub enum LadderTile {
     NcCoil,
     Horz,
     Vert,
-    BR,
-    BL,
-    UR,
-    UL,
+    LeftDown,
+    LeftUp,
+    RightDown,
+    RightUp,
     T000,
     T090,
     T180,
@@ -30,26 +31,29 @@ impl From<usize> for LadderTile {
 }
 
 impl LadderTile {
-    fn texture_filename(&self) -> &'static str {
+    fn path_string(&self) -> String {
         match self {
-            Self::Empty => "Empty",
-            Self::NoContact => "NO-Contact",
-            Self::NcContact => "NC-Contact",
-            Self::NoCoil => "NO-Coil",
-            Self::NcCoil => "NC-Coil",
-            Self::Horz => "Horz",
-            Self::Vert => "Vert",
-            Self::BR => "BR",
-            Self::BL => "BL",
-            Self::UR => "UR",
-            Self::UL => "UL",
-            Self::T000 => "T-000",
-            Self::T090 => "T-090",
-            Self::T180 => "T-180",
-            Self::T270 => "T-270",
-            Self::Cross => "Cross",
-            Self::_Length => unreachable!(),
-        }
+            LadderTile::Empty => "",
+            LadderTile::NoContact => "M 40,32 H 64 M 40,12 V 52 M 24,12 V 52 M 0,32 H 24",
+            LadderTile::NcContact => "M 44,16 L 20,48 M 24,32 H 0 M 40,32 H 64 M 40,12 V 52 M 24,12 V 52",
+            LadderTile::NoCoil => "M 48,32 H 64 M 16,32 H 0 M 36,48 A 16.67,16.67 0 0 1 36,16 M 28,16 A 16.67,16.67 0 0 1 28,48",
+            LadderTile::NcCoil => "M 44,16 L 20,48 M 36,48 A 16.67,16.67 0 0 1 36,16 M 28,16 A 16.67,16.67 0 0 1 28,48 M 64,32 H 48 M 0,32 H 16",
+            LadderTile::Horz => "M 0,32 H 64",
+            LadderTile::Vert => "M 32,0 V 64",
+            LadderTile::LeftDown => "M 0,32 H 32 V 64",
+            LadderTile::LeftUp => "M 0,32 H 32 V 0",
+            LadderTile::RightDown => "M 64,32 H 32 V 64",
+            LadderTile::RightUp => "M 64,32 H 32 V 0",
+            LadderTile::T000 => "M 0,32 H 64 M 32,32 V 64",
+            LadderTile::T090 => "M 32,0 V 64 M 32,32 H 64",
+            LadderTile::T180 => "M 0,32 H 64 M 32,32 V 0",
+            LadderTile::T270 => "M 32,0 V 64 M 32,32 H 0",
+            LadderTile::Cross => "M 0,32 H 64 M 32,0 V 64",
+            LadderTile::_Length => unreachable!(),
+        /*
+            LadderTile::_Length => todo!(),
+        */
+        }.into()
     }
 }
 
@@ -58,7 +62,6 @@ pub struct LadderTileMap {
     //TODO Rect, Vec2 or use tiles length?
     width: usize,
     height: usize,
-    tile_images: Vec<Handle<Image>>,
     tiles: Vec<Vec<Entity>>,
 }
 
@@ -76,31 +79,10 @@ impl LadderTileMap {
         LadderTileMap {
             width,
             height,
-            tile_images: default(),
             tiles: default(),
         }
     }
 
-    pub fn load_tile_images(&mut self, asset_server: &Res<AssetServer>) {
-        self.tile_images = (0..LadderTile::_Length as usize)
-            .map(|tile_variant| LadderTile::from(tile_variant))
-            .map(|tile| tile.texture_filename())
-            .map(|tile_filename| format!("./textures/{tile_filename}.png"))
-            .map(|full_path| asset_server.load(full_path).into())
-            .collect();
-    }
-
-    /*
-    fn load_tile_images_2(&mut self, asset_server: &Res<AssetServer>) {
-        let tile_length = LadderTile::_Length as usize;
-        let tile_range = 0..tile_length;
-        for tile_variant in tile_range {
-            let tile: LadderTile = tile_variant.into();
-            let full_path = format!("./textures/{}.png", tile.texture_filename());
-            self.tile_images.push(asset_server.load(full_path).into());
-        }
-    }
-    */
     #[allow(dead_code)]
     pub fn apply_pos_fn(
         &self,
@@ -116,13 +98,15 @@ impl LadderTileMap {
     }
 }
 
-pub fn ladder_image_update_system(
-    tilemap_query: Query<&LadderTileMap>, //TODO Opt, maybe store the entire handle vec in each tile? :(
-    mut tile_query: Query<(&LadderTile, &mut Handle<Image>, &Parent), Changed<LadderTile>>,
+pub fn ladder_path_update_system(
+    mut tile_query: Query<(&LadderTile, &mut Path), Changed<LadderTile>>,
 ) {
-    for (tile, mut image_handle, parent) in tile_query.iter_mut() {
-        let tilemap = tilemap_query.get(parent.get()).unwrap();
-        *image_handle = tilemap.tile_images[tile.clone() as usize].clone();
+    for (tile, mut path) in tile_query.iter_mut() {
+        *path = GeometryBuilder::build_as(&shapes::SvgPathShape {
+            svg_path_string: tile.clone().path_string(),
+            //svg_doc_size_in_px: Vec2::ZERO, //Vec2::new(64.0, 64.0),
+            svg_doc_size_in_px: Vec2::Y * 128.0, //TODO TEMP Attempt y offset
+        })
     }
 }
 
@@ -148,7 +132,6 @@ pub fn ladder_mouse_system(
     camera_query: Query<(&Camera, &GlobalTransform), With<Camera2d>>,
     mouse_buttons: Res<Input<MouseButton>>,
     tilemap_query: Query<(&LadderTileMap, &Transform)>,
-    textures: Res<Assets<Image>>,
     mut tile_query: Query<&mut LadderTile>,
 ) {
     let window = window_query.single();
@@ -158,25 +141,30 @@ pub fn ladder_mouse_system(
     let Some(cursor_world_position) = camera.viewport_to_world_2d(camera_transform, cursor_viewport_position) else { return; };
 
     for (tilemap, tilemap_transform) in tilemap_query.iter() {
-        let empty_texture = textures.get(&tilemap.tile_images[LadderTile::Empty as usize]).unwrap();
+        if tilemap.tiles.is_empty() { return; };
+        let tile_size = Vec2::splat(64.0);
 
         let delta = cursor_world_position - tilemap_transform.translation.truncate();
-        let tilemap_pixel_size = Vec2::new(tilemap.width as f32, tilemap.height as f32) * empty_texture.size();
+        let tilemap_pixel_size = Vec2::new(tilemap.width as f32, tilemap.height as f32) * tile_size;
 
         let tilemap_position = tilemap_transform.translation.truncate();
         let tilemap_rect = Rect::from_corners(tilemap_position, tilemap_position + tilemap_pixel_size);
         if !tilemap_rect.contains(cursor_world_position) { continue; };
 
-        let cursor_tile_x = (delta.x / empty_texture.size().x) as usize;
-        let cursor_tile_y = (delta.y / empty_texture.size().y) as usize;
+        let cursor_tile_x = (delta.x / tile_size.x) as usize;
+        let cursor_tile_y = (delta.y / tile_size.y) as usize;
 
         let tile_entity = tilemap.tiles[cursor_tile_x][cursor_tile_y];
-        let mut tile = tile_query.get_mut(tile_entity).unwrap();
-
+        let Ok(mut tile) = tile_query.get_mut(tile_entity) else {
+            //TODO Fix
+            dbg!("FIX ME:", tile_entity, cursor_tile_x, cursor_tile_y);
+            return;
+        };
         if mouse_buttons.just_pressed(MouseButton::Left) {
             let new_index = (*tile as usize + 1) % LadderTile::_Length as usize; //TODO Unuglify
             *tile = new_index.into();
         }
+
     }
 }
 
@@ -193,7 +181,6 @@ pub fn test_clear_tilemap_system(
                 (_, pos, size) if pos.0 == 0 || pos.0 == size.0-1 => LadderTile::Vert,
                 (_, _, _) => LadderTile::Empty,
             }
-
         }, &mut tile_query);
     }
 }
@@ -201,35 +188,49 @@ pub fn test_clear_tilemap_system(
 pub fn ladder_init_system(
     mut commands: Commands,
     mut tilemap_query: Query<(&mut LadderTileMap, Entity), Added<LadderTileMap>>,
-    textures: Res<Assets<Image>>,
 ) {
-    let empty_tile = LadderTile::default();
-    let empty_texture_index = empty_tile.clone() as usize;
     for (mut tilemap, tilemap_entity) in tilemap_query.iter_mut() {
-        let empty_texture_handle = tilemap.tile_images[empty_texture_index].clone();
-        let texture = textures.get(&empty_texture_handle).unwrap();
-        let tile_size = texture.size();
-        commands.entity(tilemap_entity).with_children(|parent_tilemap| {
+        let tile_size = Vec2::splat(64.0);
+        commands.entity(tilemap_entity)
+            .with_children(|parent_tilemap| {
             tilemap.tiles =
                 (0..tilemap.width).map(|x| {
                     (0..tilemap.height).map(|y| {
                         parent_tilemap.spawn((
                             Name::new(format!("Tile ({x},{y})")),
-                            empty_tile.clone(),
-                            SpriteBundle {
-                                texture: empty_texture_handle.clone(),
-                                sprite: Sprite {
-                                    anchor: bevy::sprite::Anchor::BottomLeft, //TODO Different anchors
-                                    ..default()
-                                },
+                            LadderTile::default(),
+                            ShapeBundle {
                                 transform: Transform::from_translation(Vec3::new(
                                     (x as f32)*tile_size.x,
                                     (y as f32)*tile_size.y, //TODO Reverse Y
-                                    0.0,
-                                )),
+                                    1.0,
+                                )).with_scale(Vec3::splat(1.0)),
+                                path: GeometryBuilder::build_as(&shapes::SvgPathShape {
+                                    svg_path_string: LadderTile::default().path_string(),
+                                    //svg_doc_size_in_px: Vec2::ZERO, //Vec2::new(64.0, 64.0),
+                                    svg_doc_size_in_px: Vec2::Y * 128.0, //TODO TEMP Attempt y offset
+                                }),
                                 ..default()
                             },
-                        )).id()
+                            Stroke::new(Color::BLACK, 2.0),
+                        ))
+                            /*
+                        .with_children(|parent_laddertile| {
+                            parent_laddertile.spawn((
+                                Text2dBundle {
+                                    text: Text::from_section("C0", TextStyle {
+                                        font_size: 24.0,
+                                        color: Color::WHITE,
+                                        ..default()
+                                    }).with_alignment(TextAlignment::Center),
+                                    text_anchor: bevy::sprite::Anchor::Center,
+                                    transform: Transform::from_xyz(0.0, 48.0, 1.0),
+                                    ..default()
+                                },
+                            ));
+                        })
+                            */
+                        .id()
                     }).collect()
                 }).collect()
             ;
